+++ source/mso-api-request/ChangeLog.md
@@ -1,5 +1,9 @@
 # ChangeLog for mso-api-request
 
+## Version 0.0.7 (2016-02-21)
+  - add option to set custom certificate file for peer verification, similar
+    to curl's --cacert option
+
 ## Version 0.0.6 (2016-01-25)
   - fix implementation of upload/scan capability